just trying some rt validation stuff out
git-svn-id: https://overture.svn.sourceforge.net/svnroot/overture/branches/vdmj-rtvalidation@2598 2b2d4d09-c93c-414a-ad04-6e973f5c1cff

+++ src/main/java/org/overturetool/vdmj/messages/rtlog/RTDeployObjectMessage.java
@@ -28,9 +28,9 @@ public class RTDeployObjectMessage extends RTArchitectureMessage
 		return this.cpu;
 	}
 	
-	public Long getObjRef()
+	public Integer getObjRef()
 	{
-		return Long.valueOf(this.object.objectReference);
+		return this.object.objectReference;
 	}
 
 }
